sicora-be-python/.docker-build/requirements.txt
# SICORA Backend Python Stack - Dependencias Consolidadas
# Optimizado para Python 3.13 - Versiones estables y compatibles

# Core FastAPI - Última versión estable
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1

# Base de datos - Últimas versiones estables
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
aiosqlite==0.20.0
greenlet==3.1.1

# Autenticación y seguridad
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
email-validator==2.2.0
cryptography==44.0.0

# HTTP y comunicación
httpx==0.28.1
python-multipart==0.0.18
redis==5.2.0
requests==2.32.3

# Testing - Versiones más recientes
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==33.3.0
pytest-mock==3.14.0

# Utilidades adicionales
python-dotenv==1.0.1
structlog==24.1.0
rich==13.9.4
click==8.1.8
typer==0.15.1

# IA y ML (sin ChromaDB por incompatibilidad con Python 3.13)
openai==1.58.1
# chromadb - Temporalmente excluido por problemas de compilación con Python 3.13
# Alternativas: pinecone-client, weaviate-client, qdrant-client

# Formateo y linting - Últimas versiones
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0

# Monitoreo y métricas
prometheus-client==0.21.1

# Documentación
mkdocs==1.6.1
mkdocs-material==9.5.48
